chore: small refactoring for train.py, folder structure, and config
