+++ src/merge_adapter.py
